requirements.txt
Flask==1.1.2
pytesseract==0.3.7
Pillow==8.1.0
gunicorn==20.0.4
